# docs/api/swagger.json
{
  "swagger": "2.0",
  "info": {},
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/graphs/versions": {
      "post": {
        "tags": [
          "graphs"
        ],
        "operationId": "createGraphVersion",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/GraphVersion"
            }
          },
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GraphVersion"
            },
            "headers": {}
          }
        }
      }
    },
    "/graphs/versions/{id}": {
      "get": {
        "tags": [
          "graphs"
        ],
        "operationId": "getGraphVersion",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GraphVersion"
            },
            "headers": {}
          }
        }
      }
    },
    "/graphs/{name}": {
      "get": {
        "tags": [
          "graphs"
        ],
        "operationId": "getGraph",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Graph"
            },
            "headers": {}
          }
        }
      },
      "post": {
        "tags": [
          "graphs"
        ],
        "operationId": "createGraph",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Graph"
            },
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GraphVersion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "tags": {
          "type": "object",
          "readOnly": true,
          "additionalProperties": {
            "$ref": "#/definitions/Tag"
          }
        },
        "structureVersionId": {
          "type": "string",
          "readOnly": true
        },
        "reference": {
          "type": "string",
          "readOnly": true
        },
        "graphId": {
          "type": "string",
          "readOnly": true
        },
        "edgeVersionIds": {
          "type": "array",
          "readOnly": true,
          "items": {
            "type": "string"
          }
        },
        "parameters": {
          "type": "object",
          "readOnly": true,
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "Graph": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string",
          "readOnly": true
        }
      }
    },
...
    "Tag": {
      "type": "object",
      "properties": {
        "versionId": {
          "type": "string",
          "readOnly": true
        },
        "key": {
          "type": "string",
          "readOnly": true
        },
        "value": {
          "type": "object",
          "readOnly": true
        },
        "valueType": {
          "type": "string",
          "readOnly": true,
          "enum": [
            "STRING",
            "INTEGER",
            "BOOLEAN"
          ]
        }
      }
    },
...
  }
}